api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
onnxruntime-gpu==1.20.1
pillow==11.0.0
numpy==1.26.4
python-multipart==0.0.12
